+++ app/build.gradle
@@ -37,6 +37,7 @@ dependencies {
     compile 'com.google.code.gson:gson:2.7'
     compile 'com.jjoe64:graphview:4.2.1'
     compile 'com.firebase:firebase-jobdispatcher:0.8.4'
+    compile 'com.android.support.constraint:constraint-layout:1.0.2'
     testCompile 'junit:junit:4.12'
     testCompile 'org.mockito:mockito-all:1.10.19'
     testCompile 'org.robolectric:robolectric:3.0'
