+++ app/src/main/AndroidManifest.xml
@@ -39,6 +39,7 @@
                 <action android:name="com.firebase.jobdispatcher.ACTION_EXECUTE"/>
             </intent-filter>
         </service>
+        <activity android:name=".MainContentActivity" />
     </application>
 
 </manifest>
