+++ app/src/main/res/layout/activity_home_content.xml
@@ -14,40 +14,38 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="0.27">
+
         <ScrollView
             android:layout_width="wrap_content"
             android:layout_height="match_parent">
+
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:gravity="center_horizontal"
                 android:orientation="vertical">
+
                 <LinearLayout
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:weightSum="3">
+                    android:weightSum="2">
 
-                    <Button
-                        android:id="@+id/fake_data_button"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:text="Fake Data"
-                        android:layout_weight="1"/>
                     <Button
                         android:id="@+id/add_point"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
-                        android:text="Add"
+                        android:layout_weight="1"
                         android:backgroundTintMode="@android:color/transparent"
-                        android:layout_weight="1"/>
+                        android:text="Add" />
+
                     <Button
                         android:id="@+id/delete_points"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"
+                        android:layout_weight="1"
                         android:backgroundTint="#FF0000"
-                        android:textSize="12sp"
                         android:text="Remove User Data"
-                        android:layout_weight="1"/>
+                        android:textSize="12sp" />
                 </LinearLayout>
 
                 <Button
@@ -65,7 +63,7 @@
                 <com.jjoe64.graphview.GraphView
                     android:id="@+id/graph"
                     android:layout_width="match_parent"
-                    android:layout_height="400dp"/>
+                    android:layout_height="400dp" />
 
                 <!--
                 <android.support.design.widget.BottomNavigationView
